R6 Class representing two datasets to be linked.

Has information on the two datasets, and the details of how to link them.

Methods

Public methods

Hismatch$new()

Hismatch$print()

Hismatch$addLinkingVariables()

Hismatch$setNames()

Hismatch$createLinkingData()

Hismatch$createBlocks()

Hismatch$setupData()

Hismatch$fuzzyMatch()

Hismatch$executeLinking()

Hismatch$mergeBackInData()

Hismatch$vectorReorder()

Hismatch$runMatching()

Hismatch$getStatistics()

Hismatch$getMatchingRatesByVariable()

Hismatch$getMatchingRatesByVariable_DT()

Hismatch$plotMatchingByGroup()

Hismatch$clone()

Method new()

Usage

Hismatch$new(
  data1 = NA,
  data2 = NA,
  firstname = NA,
  surname = NA,
  blocks = NA,
  dist_thr = 0.75,
  rel_thr = NA,
  max_block_size = 50000,
  letters = 1
)

Method print()

Usage

Hismatch$print()

Method addLinkingVariables()

Usage

Hismatch$addLinkingVariables(data_input)

Method setNames()

Usage

Hismatch$setNames(data_input, suffix)

Method createLinkingData()

Usage

Hismatch$createLinkingData(data_input)

Method createBlocks()

Usage

Hismatch$createBlocks(data)

Method setupData()

Usage

Hismatch$setupData(data_input, suffix)

Method fuzzyMatch()

Usage

Hismatch$fuzzyMatch(data1, data2, block_number)

Method executeLinking()

Usage

Hismatch$executeLinking(data1, data2, blocks_data)

Method mergeBackInData()

Usage

Hismatch$mergeBackInData(merged_data, master1, master2)

Method vectorReorder()

Usage

Hismatch$vectorReorder(dataset)

Method runMatching()

Usage

Hismatch$runMatching()

Method getStatistics()

Usage

Hismatch$getStatistics()

Method getMatchingRatesByVariable()

Usage

Hismatch$getMatchingRatesByVariable(stat_variable, data)

Method getMatchingRatesByVariable_DT()

Usage

Hismatch$getMatchingRatesByVariable_DT(variable_list, data)

Method plotMatchingByGroup()

Usage

Hismatch$plotMatchingByGroup(type_restriction)

Method clone()

The objects of this class are cloneable with this method.

Usage

Hismatch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.